Separate Leaderboard and LeaderboardHandler
Also moves them to their own package
