+++ even-more-fish-plugin/src/main/java/com/oheers/fish/competition/Competition.java
@@ -7,6 +7,7 @@ import com.oheers.fish.FishUtils;
 import com.oheers.fish.api.EMFCompetitionEndEvent;
 import com.oheers.fish.api.EMFCompetitionStartEvent;
 import com.oheers.fish.api.reward.Reward;
+import com.oheers.fish.competition.leaderboard.Leaderboard;
 import com.oheers.fish.config.CompetitionConfig;
 import com.oheers.fish.config.MainConfig;
 import com.oheers.fish.config.messages.ConfigMessage;
